backend/requirements.txt
fastapi
uvicorn
supabase
google-generativeai
openai
python-dotenv
pydantic
